// src/generated/models/Dga_ict_ai_summariesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dga_ict_ai_summariesdga_role_context = {
  1: 'Respondent',
  2: 'Reviewer',
  3: 'Approver',
  4: 'SME',
  5: 'Strategy',
  6: 'StrategyDirector',
  7: 'All'
} as const;
export type Dga_ict_ai_summariesdga_role_context = keyof typeof Dga_ict_ai_summariesdga_role_context;
export const Dga_ict_ai_summariesdga_summary_category = {
  1: 'BudgetOverview',
  2: 'EntityPortfolioSummary',
  3: 'CycleSummary',
  4: 'StrategicAlignementSummary',
  5: 'SMEReviewSummary',
  6: 'QualityCheckSummary',
  7: 'FinalApprovalSummary',
  8: 'CumulativeDocumentSummary'
} as const;
export type Dga_ict_ai_summariesdga_summary_category = keyof typeof Dga_ict_ai_summariesdga_summary_category;
export const Dga_ict_ai_summariesdga_summary_stage = {
  1: 'Planning',
  2: 'DGEReview',
  3: 'Allocation',
  4: 'Utilization'
} as const;
export type Dga_ict_ai_summariesdga_summary_stage = keyof typeof Dga_ict_ai_summariesdga_summary_stage;
export const Dga_ict_ai_summariesdga_summary_type = {
  1: 'ICTBudget',
  2: 'BudgetInstance',
  3: 'Cycle',
  4: 'StrategicPriority'
} as const;
export type Dga_ict_ai_summariesdga_summary_type = keyof typeof Dga_ict_ai_summariesdga_summary_type;
export const Dga_ict_ai_summariesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dga_ict_ai_summariesstatecode = keyof typeof Dga_ict_ai_summariesstatecode;
export const Dga_ict_ai_summariesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dga_ict_ai_summariesstatuscode = keyof typeof Dga_ict_ai_summariesstatuscode;

export interface Dga_ict_ai_summariesBase {
  dga_expiration_reason?: string;
  dga_expired_on?: string;
  dga_ict_ai_summaryid: string;
  dga_is_valid?: boolean;
  dga_model_version?: string;
  dga_name?: string;
  "dga_ReferenceRecordId@odata.bind"?: string;
  dga_referencerecordidtype?: string;
  dga_response_json?: string;
  dga_response_time?: number;
  dga_role_context?: Dga_ict_ai_summariesdga_role_context;
  dga_summary_category?: Dga_ict_ai_summariesdga_summary_category;
  dga_summary_stage?: Dga_ict_ai_summariesdga_summary_stage;
  dga_summary_type?: Dga_ict_ai_summariesdga_summary_type;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "OwningBusinessUnit@odata.bind"?: string;
  statecode: Dga_ict_ai_summariesstatecode;
  statuscode?: Dga_ict_ai_summariesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dga_ict_ai_summaries extends Dga_ict_ai_summariesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dga_is_validname?: string;
  dga_referencerecordidname?: string;
  dga_referencerecordidyominame?: string;
  dga_role_contextname?: string;
  dga_summary_categoryname?: string;
  dga_summary_stagename?: string;
  dga_summary_typename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dga_referencerecordid?: object;
  _dga_referencerecordid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}